# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile returned."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile.",
        "tags": [
          "users"
        ]
      }
    },
    "/users/devices": {
      "post": {
        "operationId": "UsersController_registerDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device registered successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a device for push notifications and E2E keys.",
        "tags": [
          "users"
        ]
      }
    },
...
    "/matches/{matchId}/video-calls": {
      "post": {
        "operationId": "VideoCallsController_scheduleCall",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleVideoCallDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule a video verification call",
        "tags": [
          "video-calls"
        ]
      }
    },
    "/matches/{matchId}/video-calls/{callId}": {
      "patch": {
        "operationId": "VideoCallsController_updateStatus",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoCallStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update video call status",
        "tags": [
          "video-calls"
        ]
      }
    },
    "/success-stories": {
      "get": {
        "operationId": "SuccessStoriesController_getPublishedStories",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SuccessStories"
        ]
      },
      "post": {
        "operationId": "SuccessStoriesController_submitStory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSuccessStoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SuccessStories"
        ]
      }
    },
    "/success-stories/{id}/approve": {
      "post": {
        "operationId": "SuccessStoriesController_approveStory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SuccessStories"
        ]
      }
    }
  },
  "info": {
    "title": "Lib le Lib API",
    "description": "API documentation for the Lib le Lib application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "description": "Device platform: 'ios' or 'android'.",
            "example": "android"
          },
          "pushToken": {
            "type": "string",
            "description": "FCM/APNs push notification token."
          },
          "publicKey": {
            "type": "string",
            "description": "E2E key-exchange public key for this device."
          }
        },
        "required": [
          "platform"
        ]
      },
...
      "ScheduleVideoCallDto": {
        "type": "object",
        "properties": {
          "scheduledAt": {
            "type": "string",
            "description": "Scheduled time for the video call"
          }
        }
      },
      "UpdateVideoCallStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "completed",
              "canceled"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateSuccessStoryDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}